FEAT: Count notion database rows (#121)
Half-Closes (https://github.com/mazipan/tanyaaja/issues/88)

## Description

There is no out of the box that Notion provided to get the Database count. The closest endpoint to get the Rows count is [Query a Database](https://developers.notion.com/reference/post-database-query).
This is not optimal since we need to do a loop, but I cannot find any other solution. So if it's accepted, in the future we need to implement a Cache.

```
GET {{base_url}}/api/statistics

{
    "message": "Public statistics",
    "data": {
        "usersCount": 128
    }
}
```